Add placeholder drawers
